+++ include/settings.h
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/28 08:45:22 by freddy            #+#    #+#             */
-/*   Updated: 2024/06/27 19:20:14 by freddy           ###   ########.fr       */
+/*   Updated: 2024/06/27 21:25:31 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -78,6 +78,10 @@
 # define BLIGHT_MINIUM_ROTATION_DEG 45
 # define BLIGHT_MAX_SPAWN_DISTANCE 10
 # define BLIGHT_SPAWN_RETURNING_DEG_VARIATION 60
+
+// animations
+
+# define BLIGHT_DEATH_ANIMATION_FRAMES 30
 
 // --- Minimap Settings